dwh_project/requirements.txt

Django==4.2.6
djangorestframework==3.14.0
matplotlib==3.8.0

